+++ OneBlink.SDK/models/FormsAppWelcomeEmail.cs
@@ -1,0 +1,8 @@
+namespace OneBlink.SDK.Model
+{
+    public class FormsAppWelcomeEmail
+    {
+        public string body { get; set; }
+        public string subject { get; set; }
+    }
+}